requirements/requirements-base.txt
flashtext==2.7
gevent==23.9.1
diskcache==5.6.3
asgiref==3.3.4
certifi
zhconv==1.4.3
click==8.0.1
fastapi==0.105.0
h11==0.12.0
pydantic==1.10.11
starlette==0.27.0
typing-extensions
uvicorn==0.14.0
uvloop==0.19.0
hypercorn==0.14.3
user-agents==2.2.0
singledispatchmethod;python_version<"3.8"
pyyaml==6.0
pycryptodome==3.19.1
loguru==0.7.2
transformers==4.38.2
emoji==2.9.0
opencc==1.1.6
pandas==2.0.3
langchain-text-splitters==0.2.0